Remove glBegin/glEnd almost everywhere.
git-svn-id: svn://svn.code.sf.net/p/opensludge/code@473 27a91dd4-ec73-4ff0-838b-64dfa05f697e

+++ source/Engine/backdrop.cpp
@@ -45,8 +45,8 @@ GLfloat backdropTexH = 1.0;
 texture lightMap;
 
 GLuint snapshotTextureName = 0;
-double snapTexW = 1.0;
-double snapTexH = 1.0;
+float snapTexW = 1.0;
+float snapTexH = 1.0;
 
 int lightMapMode = LIGHTMAPMODE_PIXEL;
